authroutes_service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
python-dotenv
httpx
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
redis
pydantic
gunicorn
pymysql
pydantic[email]
cryptography
